pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rtp_llm"
version = "0.1.0"
description = "simple building blocks for creating real-time agents"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "XvKuoMing", email = "kamasalyamov@gmail.com"}
]
license = {text = "MIT"}
keywords = ["rtp", "voice", "ai", "real-time", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi==0.115.12",
    "google-genai==1.14.0",
    "jinja2==3.1.6",
    "numpy>=1.22",
    "openai==1.76.0",
    "pydantic==2.11.3",
    "scipy>=1.10",
    "setuptools==80.0.1",
    "silero-vad>=5.1.2",
    "soundfile==0.13.1",
    "uvicorn==0.34.2",
    "webrtcvad==2.0.10",
    # Conditional dependencies using environment markers # audiopp was deprecated in python 3.13
    "audioop-lts; python_version>='3.13'",
    "librosa>=0.11.0",
    "resampy>=0.4.3",
    "python-dotenv>=1.0.0",
]

# Optional dependencies (extras)
[project.optional-dependencies]
# if user wants to run tests locally
tests = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "httpx>=0.28.1",
    "python-dotenv==1.1.0"
]

# CLI entry points
[project.scripts]
serve-rtp = "rtp_llm.entrypoints.singleton_api:main"

# Package URLs
[project.urls]
"Homepage" = "https://github.com/XvKuoMing/rtp-llm"
"Bug Reports" = "https://github.com/XvKuoMing/rtp-llm/issues"
"Source" = "https://github.com/XvKuoMing/rtp-llm.git"

# setuptools configuration
[tool.setuptools]
package-dir = {"rtp_llm" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.setuptools.package-data]
rtp_llm = ["*.md"]

